wallet 관리 핵심 기능 마무리

+++ src/main/java/com/payment/wallet/wallet/domain/Item.java
@@ -14,7 +14,7 @@ import java.math.BigDecimal;
 @SuperBuilder
 public class Item {
 
-    private Long amount;
+    private int amount;
     private String orderId;
     private Long referenceId;
     private ReferenceType referenceType;
