+++ src/main/java/com/payment/wallet/wallet/domain/PaymentOrderDTO.java
@@ -13,11 +13,10 @@ public class PaymentOrderDTO extends Item{
 
     private Long id;
     private Long sellerId;
-    private Long amount;
+    private int amount;
     private String orderId;
 
-
-    public PaymentOrderDTO(Long amount, String orderId, Long referenceId, ReferenceType referenceType, Long id, Long sellerId) {
+    public PaymentOrderDTO(int amount, String orderId, Long referenceId, ReferenceType referenceType, Long id, Long sellerId) {
         super(amount, orderId, referenceId, referenceType);
         this.id = id;
         this.sellerId = sellerId;
